# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StePanel API
  version: 0.7.0
  description: |
    Authenticated server operations, service inventory, and migration API.
    Mutating requests require the `stepanel_session` cookie and the matching
    `stepanel_csrf` value in the `X-CSRF-Token` header or `csrf` form field.
    Responses include an `X-Request-ID` correlation identifier.
servers:
  - url: http://127.0.0.1:8090
security:
  - sessionCookie: []
paths:
...
  /api/tasks/{site}:
    get:
      summary: List bounded systemd-timer scheduled tasks for a site
      parameters: [{name: site, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Scheduled task definitions}}
  /api/tasks/{site}/{name}:
    put:
      summary: Create or replace a site-identity scheduled task
      parameters: [{name: site, in: path, required: true, schema: {type: string}}, {name: name, in: path, required: true, schema: {type: string}}]
      responses: {'202': {description: Systemd service and timer applied}}
    delete:
      summary: Remove a scheduled task service and timer
      parameters: [{name: site, in: path, required: true, schema: {type: string}}, {name: name, in: path, required: true, schema: {type: string}}]
      responses: {'204': {description: Scheduled task removed}}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: stepanel_session